refactor: DetailedStation***을 StationDetails*** 으로 수정
[#363]
