--- frontend/src/components/ui/DetailedStationInfo/index.tsx
@@ -1,3 +1,0 @@
-import DetailedStationInfo from './DetailedStationInfo';
-
-export default DetailedStationInfo;
