+++ frontend/src/components/ui/StationDetailsWindow/ChargerCard.tsx
