+++ frontend/src/components/ui/StationDetailsWindow/StationDetailsView.stories.tsx
@@ -1,11 +1,11 @@
 import type { Meta } from '@storybook/react';
 
-import type { DetailedStationProps } from '@ui/DetailedStationInfo/DetailedStation';
-import DetailedStation from '@ui/DetailedStationInfo/DetailedStation';
+import type { DetailedStationProps } from '@ui/StationDetailsWindow/StationDetailsView';
+import StationDetailsView from '@ui/StationDetailsWindow/StationDetailsView';
 
 const meta = {
-  title: 'UI/DetailedStation',
-  component: DetailedStation,
+  title: 'UI/StationDetailsView',
+  component: StationDetailsView,
   tags: ['autodocs'],
   args: {
     station: {
@@ -72,10 +72,10 @@ const meta = {
       description: '충전소 데이터를 수정할 수 있습니다.',
     },
   },
-} satisfies Meta<typeof DetailedStation>;
+} satisfies Meta<typeof StationDetailsView>;
 
 export default meta;
 
 export const Default = (args: DetailedStationProps) => {
-  return <DetailedStation {...args} />;
+  return <StationDetailsView {...args} />;
 };
