+++ frontend/src/components/ui/StationDetailsWindow/StationInformation.tsx
