+++ frontend/src/components/ui/StationDetailsWindow/congestion/CongestionStatistics.tsx
