+++ frontend/src/components/ui/StationDetailsWindow/index.tsx
@@ -1,0 +1,3 @@
+import StationDetailsWindow from './StationDetailsWindow';
+
+export default StationDetailsWindow;
