+++ frontend/src/components/ui/StationDetailsWindow/reports/ChargerReportConfirmation.tsx
